new update structure files
