+++ script.mjs
@@ -1,20 +1,29 @@
-import { reloadWindow, loaderHandler, setLocalStorageInfo, setupUi, createAlert, templateCard, handleLogout } from "./src/js/index.mjs";
+// imports
+import {
+  loaderHandler, setLocalStorageInfo, setupUi,
+  createAlert, templateCard, handleLogout, handleClickDeleteButton,
+  handlePagination, handleLogin, handleRegister, handleCreatePost,
+   handleAddingComment, showUserInfo, templateComment, postTemplate,
+   handleClickCard
+} from "./src/js/index.mjs";
 
-// const variable globel
-const url = "https://tarmeezacademy.com/api/v1";
-let postArray = [];
-let currentPage = 1;
-let lastPage;
-let postInfo = [];
-let user = {};
+// variable globel
+export const url = "https://tarmeezacademy.com/api/v1"; // the main url
+export let postArray = []; // for the posts in getReqeust
+export let currentPage = { value: 1 }; // for the pagination
+export let lastPage; // for the pagination
+export let postInfo = { value: [] }; // for pagination when its have data stop the pagination
+export let user = {}; // for the user info
 let updatePost = false;
-let postIdUpdate = null;
-let currentPostClick = null;
+export let idPost = { value: null }; // for the post id handle for the delete and update
+export let currentPostClick = null; // for the post click
+export let isFetching = { value: false }; // for the pagination
+
 
 /**
  * request for the api for data posts
  */
-async function getRequest(updatePost, current) {
+export async function getRequest(updatePost, current) {
   loaderHandler(true);
   const response = await axios.get(`${url}/posts?limit=5&page=${current}`)
     .then((response) => {
@@ -31,33 +40,18 @@ async function getRequest(updatePost, current) {
         posts.innerHTML = "";
       }
       postArray.map((item) => {
-        console.log(item);
         currentPostClick = item;
-
         // loop for the tags for every post
         const tags = item.tags.map((tg) => {
           return `<div> ${tg}</div>`;
         });
-
-        let title = "";
-        // check if the title is null then  remove the title else add the title
-        if (item.title == null) {
-          title = "";
-        } else {
-          title = item.title;
-        }
-
+        let title = item.title == null ? "" : item.title;
         let id = item.id;
-        // for update method
-        postIdUpdate = id;
-        const authorIdPost = item.author.id;
-        // oldEcma script code
-        let idUser;
-        if (user !== null) {
-          idUser = user.id;
-        }
+        idPost.value = id; // for the post id handle for the delete and update
+        const authorIdPost = item.author?.id;
+        let idUser = user !== null ? user.id : " "
         let conditionEdit = idUser != null && authorIdPost == idUser;
-        posts.innerHTML += templateCard(item, conditionEdit, postIdUpdate, id, title, tags);
+        posts.innerHTML += templateCard(item, conditionEdit, idPost.value, id, title, tags);
       });
     })
     .catch((error) => {
@@ -73,268 +67,34 @@ async function getRequest(updatePost, current) {
 /**
  * adding pagination for the website
  */
-//  fix
-let isFetching = false;
-window.addEventListener("scroll", () => {
-  if (
-    !isFetching &&
-    window.pageYOffset + 3000 >= document.body.offsetHeight &&
-    currentPage < lastPage &&
-    postInfo.length == 0
-  ) {
-    isFetching = true; // Set flag to true when a request is initiated
-    currentPage = currentPage + 1;
-    getRequest(false, currentPage)
-      .then(() => {
-        isFetching = false;
-      })
-      .catch(() => {
-        console.log("error happend pagination");
-      });
-  }
-});
+window.addEventListener("scroll", () => handlePagination());
 
 /**
  * for loginBtn handler and send the data
  */
-document.querySelector("#LoginBtn").addEventListener("click", () => {
-  let username = document.querySelector("#username-input-login").value;
-  let password = document.querySelector("#password-input-login").value;
-  let data = {
-    username: username,
-    password: password,
-  };
-  loaderHandler(true);
-  axios
-    .post(`${url}/login`, data)
-    .then((response) => {
-      // hide modal
-      const modal = document.getElementById("login-modal");
-      const modalInstance = bootstrap.Modal.getInstance(modal);
-      modalInstance.hide();
-      setLocalStorageInfo(
-        response.data.token,
-        JSON.stringify(response.data.user),
-      );
-      // createAlert();
-      createAlert("Login successful! Welcome back!", "success");
-      setupUi();
-      loaderHandler(false);
-    })
-    .catch((e) => {
-      createAlert(`error happend: ${e}`, "danger");
-      loaderHandler(false);
-    });
-});
+document.querySelector("#LoginBtn").addEventListener("click", () => handleLogin());
 
 /**
  * handle the logout
  */
 document.getElementById("logout-button").addEventListener("click", () => handleLogout());
 
+/**
+ * for the first time to setup the ui for the app
+ */
 setupUi();
 
 /**
- * handle register
+ * handle register 
  */
-document.getElementById("RegisterBtn").addEventListener("click", () => {
-  console.log("------- register ----------");
-  const user = document.getElementById("user-register").value;
-  const username = document.getElementById("username-register").value;
-  const email = document.getElementById("email-register").value;
-  const password = document.getElementById("password-register").value;
-  const image = document.getElementById("image-register").files[0];
-
-  const dataForm = new FormData();
-  dataForm.append("name", user);
-  dataForm.append("username", username);
-  dataForm.append("email", email);
-  dataForm.append("password", password);
-  dataForm.append("image", image);
-
-  loaderHandler(true);
-  axios
-    .post(`${url}/register`, dataForm)
-    .then((response) => {
-      console.log(response);
-      createAlert("Registration successful! now login.", "info");
-      const modal = document.getElementById("register-modal");
-      const modalInstance = bootstrap.Modal.getInstance(modal);
-      modalInstance.hide();
-      setLocalStorageInfo(response.data.token, JSON.parse(response.data.user));
-      setupUi();
-      loaderHandler(false);
-    })
-    .catch((e) => {
-      // old ecame script
-      createAlert(`error: ${e?.response?.data?.message}`, "danger");
-      loaderHandler(false);
-    });
-});
+document.getElementById("RegisterBtn").addEventListener("click", () => handleRegister());
 
 /**
  * create post for user that authorization
- * FIX: problem....
  */
-document.querySelector("#create-post-button").addEventListener("click", () => {
-  loaderHandler(true);
-  const title = document.querySelector("#title-create-post").value;
-  const body = document.querySelector("#body-create-post").value;
-  const image = document.querySelector("#image-create-post").files[0];
-  const token = localStorage.getItem("token");
-  const formData = new FormData();
-  formData.append("title", title);
-  formData.append("body", body);
-  formData.append("image", image);
-  const headers = {
-    authorization: `Bearer ${token}`,
-  };
-  axios
-    .post(`${url}/posts`, formData, { headers: headers })
-    .then((response) => {
-      console.log(response);
-      createAlert("success create a new post ... ", "success");
-      const modal = document.getElementById("create-post-modal");
-      const modalInstance = bootstrap.Modal.getInstance(modal);
-      modalInstance.hide();
-      getRequest(true);
-      loaderHandler(false);
-    })
-    .catch((e) => {
-      createAlert("error happend: " + e.request.responseText, "danger");
-      console.log(e);
-      loaderHandler(false);
-    });
-});
+document.querySelector("#create-post-button").addEventListener("click", () => handleCreatePost());
 
-/**
- * handle create an post page
- */
-export const handleClickCard = (e) => {
-  // console.log(e);
-  loaderHandler(true);
-  axios
-    .get(`${url}/posts/${e}`)
-    .then((response) => {
-      postInfo = [];
-      postInfo = response.data.data;
-      console.log("handle click card handler")
-      console.log(postInfo);
-    })
-    .then(() => {
-      let containerPost = document.querySelector(".container__posts");
-      containerPost.innerHTML = "";
-      let postIdUpdate = postInfo.id;
-      const authorIdPost = postInfo.author.id;
-      let conditionEdit = postIdUpdate != null && authorIdPost == postIdUpdate;
-      let item = postInfo;
-      const allComments = postInfo.comments.map((item) => {
-        const comm = `
-        <div class="comment pb-1" onclick={showUserInfo('${encodeURIComponent(JSON.stringify(postInfo),)}')} >
-          <div class="pt-1 flex align-items-center justify-content-center gap-3 pb-1">
-          ${typeof item.author.profile_image === "string" ?
-            `<img src="${item.author.profile_image}" alt="Avatar1" class="img-fluid my-1" style="width: 30px; height: 30px; border-radius: 50% !important;" />`
-            :
-            `<img src="images/icons8-user-48.png" alt="Avatar2" class="img-fluid my-1" style="width: 30px; height: 30px; border-radius: 50% !important;" />`
-          }                     
-          <span><strong>@${item.author.username}</strong></span>
-          </div>
-          <h5>${item.body}</h5>
-        </div>
-      `;
-        return comm;
-      });
-
-      const post = `
-      <h2 style="padding: 40px 0px 0px; color: white !important;">${postInfo.author.username} post</h2>
-      <div class="card my-5">
-        <div class="card-header">
-          <img src="${postInfo.author.profile_image}" style="width: 40px; border-radius: 50%;" />
-          <span onclick="showUserInfo('${encodeURIComponent(
-        JSON.stringify(item),
-      )}')" style="cursor: pointer;"><strong>${postInfo.author.username}</strong></span>
-            ${conditionEdit
-          ? `<button class="btn btn-danger mx-2" style="float: right;" onclick={handleClickDeleteButton('${postIdUpdate}')}>Delete</button>`
-          : ""
-        }
-            ${conditionEdit
-          ? `<button class="btn btn-primary" style="float: right;" onclick={handleClickEditButton('${encodeURIComponent(
-            JSON.stringify(item),
-          )}')}>edit</button>`
-          : ""
-        }
-        </div>
-        <div class="card-body">
-          <img src="${postInfo.image}" alt="image post" style="width: 100%; height: 500px; padding-bottom: 3px;" />
-          <h6 style="color: #777;">${postInfo.create_at}</h6>
-          <h4>${postInfo.title}</h4>
-          <p>${postInfo.body}</p>
-          <hr />
-          <div class="d-flex gap-3 justify-content-start align-items-center">
-            <a href="#">
-              <span>
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pen-fill" viewBox="0 0 16 16">
-                  <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001" />
-                </svg>
-              </span>
-              (${postInfo.comments_count}) Commits
-            </a>
-          </div>
-          <!-- Comments Section -->
-            <br />
-          <div class="comments-section">
-            <h3 style="text-decoration: underline; margin-bottom: 20px; width: 300px;">Comments</h3>
-              ${allComments}
-          </div>
-        </div>
-      <div class="input-group">
-        <input id="comment__input" type="text" class="form-control py-2" placeholder="add comment..." aria-label="Enter something..." aria-describedby="button-addon">
-        <button class="btn btn-primary" type="button" id="button-addon" onclick="handleAddingComment(${e})">Add Comment</button>
-      </div>
-      </div>
-`;
-      containerPost.innerHTML = post;
-
-    })
-    .catch((e) => {
-      console.log("error happend", e);
-      createAlert("error happend " + e, "danger");
-    });
-  loaderHandler(false);
-};
-
-/**
- * handle create comment for  the user
- */
-export const handleAddingComment = (e) => {
-  console.log("adding comment");
-  let commentValue = document.querySelector("#comment__input").value;
-  let token = localStorage.getItem("token");
-  const data = {
-    body: commentValue,
-  };
-  const headers = {
-    authorization: `Bearer ${token}`,
-  };
-  if (commentValue == "") {
-    createAlert("please enter the comment", "danger");
-    return;
-  }
-  loaderHandler(true);
-  axios
-    .post(`${url}/posts/${e}/comments`, data, { headers: headers })
-    .then((response) => {
-      // console.log(response);
-      createAlert("success adding commit to the post", "success");
-      handleClickCard(e);
-      loaderHandler(false);
-    })
-    .catch((e) => {
-      console.log("error happend", e);
-      createAlert("error happend " + e.request.response, "danger");
-      loaderHandler(false);
-    });
-};
+document.getElementById("profile-user").addEventListener("click", () => showUserInfo(null));
 
 /**
  * handle click on edit button for my post
@@ -351,201 +111,5 @@ export const handleClickEditButton = (e) => {
   modalInstance.toggle();
 };
 
-/**
- * for handle the deleting button for the post that user made
- */
-export const handleClickDeleteButton = (e, refresh) => {
-  loaderHandler(true);
-  const token = localStorage.getItem("token");
-  const headers = {
-    authorization: `Bearer ${token}`,
-  };
-  axios
-    .delete(`${url}/posts/${e}`, { headers: headers })
-    .then((response) => {
-      console.log(response);
-      getRequest(true);
-      createAlert("deleting the post successfuly ", "success");
-      // FIX handle the fix click
-      if (refresh) {
-        showUserInfo(currentPostClick);
-      }
-      loaderHandler(false);
-    })
-    .catch((e) => {
-      console.log(e);
-      createAlert("error happend in deleting", "danger");
-      loaderHandler(false);
-    });
-};
-
-/**
- * for show the user info
- */
-export const showUserInfo = async (element) => {
-  // to stop fetching data for pagination
-  isFetching = true;
-  let content = document.querySelector(".container__posts");
-  let user;
-  let userProfile;
-  loaderHandler(true);
-  if (element == null) {
-    user = JSON.parse(localStorage.getItem("user"));
-    userProfile = true;
-  } else {
-    user = JSON.parse(decodeURIComponent(element));
-    console.log(user)
-    userProfile = false;
-    let user_id = user.author.id;
-    console.log(user)
-    const request = await axios
-      .get(`${url}/users/${user_id}`)
-      .then((response) => {
-        user = response.data.data;
-      })
-      .catch((e) => console.log(e));
-  }
-  if (user == null) {
-    createAlert("no user sign in ..", "danger");
-  } else {
-    let postUser;
-    console.log(user.id);
-    loaderHandler(true);
-    const requestPost = await axios
-      .get(`${url}/users/${user.id}/posts`)
-      .then((response) => {
-        postUser = response.data.data;
-      })
-      .catch((e) => {
-        console.log("error happend", e);
-      });
-    loaderHandler(false);
-    const allPostUser = postUser.map((item) => {
-      // loop for the tags for every post
-      const tags = item.tags.map((tg) => {
-        return `<div> ${tg}</div>`;
-      });
-
-      let title = "";
-      // check if the title is null then  remove the title else add the title
-      if (item.title == null) {
-        title = "";
-      } else {
-        title = item.title;
-      }
-
-      let id = item.id;
-      // for update method
-      postIdUpdate = id;
-
-      return `
-   <div class="card" >
-                <div class="card-header">
-                     ${typeof item.author.profile_image === "string" ?
-          `<img src="${item.author.profile_image}" alt="Avatar1" class="img-fluid my-1" style="width: 30px; height: 30px; border-radius: 50% !important;" />`
-          :
-          `<img src="images/icons8-user-48.png" alt="Avatar2" class="img-fluid my-1" style="width: 30px; height: 30px; border-radius: 50% !important;" />`
-        }                     
-       <span onclick={showUserInfo('${encodeURIComponent(
-          JSON.stringify(item),
-        )}')} style="cursor: pointer;"><strong>${item.author.username
-        }</strong></span>
-                    ${userProfile
-          ? `<button class="btn btn-danger mx-2" style="float: right;" onclick={handleClickDeleteButton('${postIdUpdate}')}>Delete</button>`
-          : ""
-        }
-                    ${userProfile
-          ? `<button class="btn btn-primary" style="float: right;" onclick={handleClickEditButton('${encodeURIComponent(
-            JSON.stringify(item),
-          )}' )}>edit</button>`
-          : ""
-        }
-                </div>
-                <div class="card-body" onclick="handleClickCard(${id})">
-                    <img src=${item.image} alt="image post"
-                    style="width: 100%;height: 500px;padding-bottom: 3px;">
-                    <h6 style="color: #777;">${item.created_at}</h6>
-                    <h4>${title}</h4>
-                    <p>${item.body}</p>
-                    <hr />
-                    <div class="d-flex gap-3 justify-content-start align-items-center">
-                        <a href="#"> <span>
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pen-fill" viewBox="0 0 16 16"> <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001" />
-                        </svg> </span>
-                            (${item.comments_count})Commits
-                        </a>
-                        <span style="display: flex; gap: 5px;" class="category">
-                            ${tags}
-                        </span>
-                    </div>
-                </div>
-            </div>
-    `;
-    });
-
-    let userInfo = `
-        < section class="h-100 profile__user" style = "background-color: #f4f5f7; color: black;" >
-          <div class="container profile__user py-2 h-100 w-100">
-            <div class="row d-flex justify-content-center align-items-center">
-              <div class="col col-lg-9 mb-4 mb-lg-9">
-                <div class="card mb-3" style="border-radius: .5rem;">
-                  <div class="row g-0">
-                    <div class="col-md-4 text-center text-white"
-                      style="border-top-left-radius: .5rem; border-bottom-left-radius: .5rem;  background-image: radial-gradient(circle, #051937, #09485f, #3f7a7d, #84ac98, #d2dcbc);">
-                      ${typeof user.profile_image === "string" ?
-        `<img src="${user.profile_image}" alt="Avatar1" class="img-fluid my-5" style="width: 80px; height: 80px; border-radius: 50% !important;" />`
-        :
-        `<img src="images/icons8-user-48.png" alt="Avatar2" class="img-fluid my-5" style="width: 80px; height: 80px; border-radius: 50% !important;" />`
-      }
-                      <h5 class="text-white">${user.username}</h5>
-                      <i class="far fa-edit mb-5"></i>
-                    </div>
-                    <div class="col-md-8">
-                      <div class="card-body p-4">
-                        <h6>Information</h6>
-                        <hr class="mt-0 mb-4">
-                          <div class="row pt-1 text-white">
-                            <div class="col-6 mb-3">
-                              <h6>Email</h6>
-                              <p class="text-muted text-white" style="color: white !important;">${user.email}</p>
-                            </div>
-                            <div class="col-6 mb-3">
-                              <h6>Phone</h6>
-                              <p style="color: white !important;" class="text-muted text-white">123 456 789</p>
-                            </div>
-                          </div>
-                          <h6>Activites</h6>
-                          <hr class="mt-0 mb-4">
-                            <div class="row pt-1">
-                              <div class="col-6 mb-3">
-                                <h6>Recent Posts</h6>
-                                <p style="color: white !important;" class="text-muted">${user.posts_count}</p>
-                              </div>
-                              <div class="col-6 mb-3">
-                                <h6>Comments Counts</h6>
-                                <p style="color: white !important;" class="text-muted">${user.comments_count}</p>
-                              </div>
-                            </div>
-                          </div>
-                      </div>
-                    </div>
-                  </div>
-                  <h2 style="text-decoration: underline; font-weight: bold;"> Posts User </h2>
-                  ${allPostUser}
-                </div>
-              </div>
-            </div>
-          </section>
-      `;
-    content.innerHTML = userInfo;
-  }
-  loaderHandler(false);
-};
-
-document
-  .getElementById("profile-user")
-  .addEventListener("click", () => showUserInfo(null));
-
-
 // request for the first time
-getRequest(false, currentPage);
+getRequest(false, currentPage.value);
