+++ src/js/Logic/pagination.mjs
@@ -1,0 +1,23 @@
+import { isFetching,getRequest, postInfo, currentPage, lastPage } from "../../../script.mjs";
+import { createAlert } from "../index.mjs";
+function handlePagination() {
+if (
+    !isFetching.value &&
+    window.pageYOffset + 3000 >= document.body.offsetHeight &&
+    currentPage.value < lastPage &&
+    postInfo.value.length == 0
+  ) {
+    isFetching.value = true; // Set flag to true when a request is initiated
+    currentPage.value = currentPage.value + 1;
+    getRequest(false, currentPage.value)
+      .then(() => {
+        isFetching.value = false;
+      })
+      .catch(() => {
+        createAlert("error happend pagination", "danger");
+        console.log("error happend pagination");
+      });
+  }
+}
+
+export default handlePagination;
