+++ src/js/index.mjs
@@ -6,10 +6,23 @@ import createAlert from "./UI/createAlert.mjs";
 import createUserLoginInfoNavBar from "./UI/createUserInfoNav.mjs";
 import templateCard from "./templates/templateCard.mjs";
 import handleLogout from "./Logic/logout.mjs";
-// import handleClickDeleteButton from "./Logic/handleDeletingPost.mjs";
+import handleClickDeleteButton from "./Logic/handleDeletingPost.mjs";
+import handlePagination from "./Logic/pagination.mjs";
+import handleLogin from "./Logic/handleLogin.mjs";
+import handleRegister from "./Logic/handleRegister.mjs";
+import handleCreatePost from "./Logic/handleCreatePost.mjs";
+import handleAddingComment from "./Logic/handleAddingComment.mjs";
+import userProfilePage from "./templates/userProfilePage.mjs";
+import userInfoPostTemplate from "./templates/userInfoPostTemplate.mjs";
+import showUserInfo from "./Logic/handleShowInfoUser.mjs";
+import templateComment from "./templates/commentTemplate.mjs";
+import postTemplate from "./templates/postTemplate.mjs";
+import handleClickCard from "./Logic/handleClickCard.mjs";
 
 export {
     reloadWindow, loaderHandler, setLocalStorageInfo,
     setupUi, createAlert, createUserLoginInfoNavBar, templateCard,
-    handleLogout 
+    handleLogout ,handleClickDeleteButton, handlePagination, handleLogin,
+    handleRegister, handleCreatePost, handleAddingComment, userInfoPostTemplate, userProfilePage, showUserInfo,
+    templateComment, postTemplate, handleClickCard
 };
