+++ src/js/templates/templateCard.mjs
@@ -1,5 +1,5 @@
-import { handleClickCard, handleAddingComment,  handleClickEditButton , handleClickDeleteButton} from "../../../script.mjs";
-import { showUserInfo } from "../../../script.mjs";
+import {    handleClickEditButton } from "../../../script.mjs";
+import { handleClickDeleteButton , handleAddingComment, showUserInfo , handleClickCard } from "../index.mjs";
 
 function templateCard(item, condition, idUpdate, id, title, tags) {
   window.showUserInfo = showUserInfo;
@@ -34,13 +34,11 @@ function templateCard(item, condition, idUpdate, id, title, tags) {
         <p>${item.body}</p>
         <hr />
         <div class="d-flex gap-3 justify-content-start align-items-center">
-            <a href="#">
-             <span>
-                  <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentcolor" class="bi bi-pen-fill" viewbox="0 0 16 16"> <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059l4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056l6.854 4.854a.5.5 0 1 1-.708-.708l9.44.854a1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001" />
-                  </svg> 
-              </span>
-                  (${item.comments_count})commits
-            </a>
+        <a href="#"> <span>
+        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pen-fill" viewBox="0 0 16 16"> <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001" />
+        </svg> </span>
+            (${item.comments_count})Commits
+        </a>
             <span style="display: flex; gap: 5px;" class="category">
                 ${tags}
             </span>
